Add files via upload
Done via examples on google.

Make sure to go in the index1.html and change the links to the documents linking them to Products and Contacts pages.
Also, make sure to add the CSS and JS files as well. 

If this works, you can change the name to index.html instead of index1.